Add early return if game reset during computer thinking
Async method 'computerTurnLogic' is now checking if game is still in
progress after it is done with delay so as not to continue executing.
Before it was possible to continue the game after quickly pressing
'Start / Reset' button twice, if after the first press it was
computer's turn, and the second press came before 'Task.Delay' that is
simulating computer's thinking in 'computerTurnLogic' was completed.