Update ship sinking logic and add comments
The `MarkShipSunken` method in `Gunnery.cs` now sorts the ship squares
by row and column before passing them to `eliminator.ToEliminate`. This
sorting is required for 'ToEliminate' method to work correctly.

Before this change, if the squares of the sunken ship were hit in order
other than top to bottom or left to right, so that squares passed on to
eliminator were ordered by rows and columns by pure chance, not all
squares surrounding the sunken ship were marked as eliminated.